@@ -1,10 +1,10 @@
 name: mx_env_example
 description: Demonstrates how to use the mx_env plugin.
 publish_to: 'none'
-version: 0.0.1+1
+version: 0.1.1+11
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
